make examples using scalacheck green
Adding pow() example run with small values is needed cause there was
situation when most of tests with large input values passed but not too
large values faced with number overflow.

+++ src/main/scala-2.11/impatient/scala/exercises/Chapter2.scala
@@ -63,7 +63,7 @@ class Chapter2 {
   }
 
 
-  def pow(base: Int)(power: Int): Float = {
+  def pow(base: Int)(power: Int): Double = {
     power match {
       case 0 => 1
 
@@ -74,7 +74,7 @@ class Chapter2 {
         base * pow(base)(x - 1)
 
       case x if x > 0 && x % 2 == 0 =>
-        val tmp = pow(base)(x / 2).toInt
+        val tmp = pow(base)(x / 2)
         tmp * tmp
 
     }
